Improvements

+++ src/com/bachue/snr/biometrico/admon/persistence/ejb/dao/stateless/impl/UsuarioDAOImpl.java
@@ -64,7 +64,8 @@ public class UsuarioDAOImpl implements IUsuarioDAO {
 	public Boolean borrarUsuario(String idUsuario) {
 		EntityManager lem_entityManager = iiemf_entityFactory.getEntityManager();
 		Usuario lu_usuario = lem_entityManager.find(Usuario.class, idUsuario);
-		lem_entityManager.remove(lu_usuario);
+		if(lu_usuario != null)
+			lem_entityManager.remove(lu_usuario);
 		lem_entityManager.close();
 		return true;
 	}
